requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
autopep8==2.0.2
    # via -r requirements.in
bleach==6.0.0
    # via readme-renderer
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
colorama==0.4.6
    # via pytest
contourpy==1.0.7
    # via matplotlib
coverage[toml]==7.2.4
    # via
    #   -r requirements.in
    #   pytest-cov
cycler==0.11.0
    # via matplotlib
docutils==0.19
    # via readme-renderer
exceptiongroup==1.1.1
    # via pytest
fonttools==4.39.3
    # via matplotlib
idna==3.4
    # via requests
importlib-metadata==6.6.0
    # via
    #   keyring
    #   twine
importlib-resources==5.12.0
    # via matplotlib
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.2.3
    # via keyring
keyring==23.13.1
    # via twine
kiwisolver==1.4.4
    # via matplotlib
markdown-it-py==2.2.0
    # via rich
matplotlib==3.7.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
more-itertools==9.1.0
    # via jaraco-classes
numpy==1.24.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
packaging==23.1
    # via
    #   matplotlib
    #   pytest
pandas==2.0.1
    # via -r requirements.in
pillow==9.5.0
    # via matplotlib
pkginfo==1.9.6
    # via twine
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest-html
pycodestyle==2.10.0
    # via autopep8
pygments==2.15.1
    # via
    #   readme-renderer
    #   rich
pyparsing==3.0.9
    # via matplotlib
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
pytest-cov==4.0.0
    # via -r requirements.in
pytest-html==3.2.0
    # via -r requirements.in
pytest-metadata==2.0.4
    # via pytest-html
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
pywin32-ctypes==0.2.0
    # via keyring
readme-renderer==37.3
    # via twine
requests==2.29.0
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.10.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.3.5
    # via twine
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
tomli==2.0.1
    # via
    #   autopep8
    #   coverage
    #   pytest
twine==4.0.2
    # via -r requirements.in
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via
    #   requests
    #   twine
webencodings==0.5.1
    # via bleach
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources
